+++ core/auth/auth_type_email.go
@@ -40,7 +40,7 @@ type emailAuthImpl struct {
 	authType string
 }
 
-func (a *emailAuthImpl) signUp(authType model.AuthType, appType model.ApplicationType, appOrg model.ApplicationOrganization, creds string, params string, newCredentialID string, l *logs.Log) (string, map[string]interface{}, error) {
+func (a *emailAuthImpl) signUp(authType model.AuthType, appOrg model.ApplicationOrganization, creds string, params string, newCredentialID string, l *logs.Log) (string, map[string]interface{}, error) {
 	type signUpEmailParams struct {
 		ConfirmPassword string `json:"confirm_password"`
 	}
@@ -103,7 +103,7 @@ func (a *emailAuthImpl) signUp(authType model.AuthType, appType model.Applicatio
 
 	if verifyEmail {
 		//send verification code
-		if err = a.sendVerificationCode(email, appType.Application.Name, code, newCredentialID); err != nil {
+		if err = a.sendVerificationCode(email, appOrg.Application.Name, code, newCredentialID); err != nil {
 			return "", nil, errors.WrapErrorAction(logutils.ActionSend, "verification email", nil, err)
 		}
 	}